+++ src/basilisp/lang/seq.py
@@ -228,7 +228,9 @@ def sequence(s: Iterable[T], support_single_use: bool = False) -> ISeq[T]:
     i = iter(s)
 
     if not support_single_use and i is s:
-        raise TypeError(f"Can't create sequence out of single-use iterable object, please use iterator-seq instead. Iterable Object type: {type(s)}")
+        raise TypeError(
+            f"Can't create sequence out of single-use iterable object, please use iterator-seq instead. Iterable Object type: {type(s)}"
+        )
 
     def _next_elem() -> ISeq[T]:
         try:
@@ -244,6 +246,7 @@ def sequence(s: Iterable[T], support_single_use: bool = False) -> ISeq[T]:
 def iterator_sequence(s: Iterable[T]) -> ISeq[T]:
     """Create a Sequence from any iterable `s`."""
     return sequence(s, support_single_use=True)
+
 
 @overload
 def _seq_or_nil(s: None) -> None: ...